fix: eliminate all 9 circular references and gate against new ones in CI
Resolve the 9 circular dependencies reported by `madge --circular`
(npm run check-cycle-references) and convert the existing PR-comment
summary in test.yml into a required CI gate.

Each cycle was broken by relocating the shared symbol that creates the
back-edge, or by inverting one edge — no behavior change:

- inso db/models/types <-> db/types: move `Database` to live with the
  model types it references; db/types re-exports it.
- network.ts <-> common/render.ts: extract getOrInherit{Authentication,
  Headers} into network/inherit.ts (network re-exports for tests).
- network.ts <-> scripting-env objects (cross-package): import
  ExecutionOption/RequestContext from their leaf files, and move
  sendCurlAndWriteTimeline* types into network/network.types.ts.
- import-export.tsx <-> export-requests-modal.tsx: extract the pure
  export functions into import-export.utils.tsx (re-exported).
- window-utils/plugin-window/prompt-bridge: extract browserWindows +
  getMainWindow into main-window-registry.ts.
- response-viewer <-> response-multipart-viewer (recursion): inject
  ResponseViewer into the multipart viewer via a prop; props type moved
  to response-viewer.types.ts.
- tag-editor-arg-sub-form <-> external-vault-form: extract
  ArgConfigFormProps to a sibling types file.
- tab-list <-> tab: move TAB_CONTEXT_MENU_COMMAND enum into tab.tsx.

CI gate: the "Check Circular References" step now fails (core.setFailed)
when a PR introduces any cycle not present in the base branch — gated on
new cycles rather than net count, so swapping one cycle for another still
fails. The PR comment is still posted.

Validated: madge 0 cycles, type-check, lint (0 errors), network +
scripting-environment unit tests pass.
