+++ packages/insomnia-inso/src/db/types.ts
@@ -1,30 +1,6 @@
-import type { CaCertificate, ClientCertificate, CloudProviderCredential, CookieJar, Settings } from 'insomnia-data';
+import type { Database } from './models/types';
 
-import type {
-  ApiSpec,
-  BaseModel,
-  Environment,
-  UnitTest,
-  UnitTestSuite,
-  Workspace,
-  WorkspaceMeta,
-} from './models/types';
-
-export interface Database {
-  ApiSpec: ApiSpec[];
-  Environment: Environment[];
-  Request: BaseModel[];
-  RequestGroup: BaseModel[];
-  Workspace: Workspace[];
-  WorkspaceMeta: WorkspaceMeta[];
-  UnitTestSuite: UnitTestSuite[];
-  UnitTest: UnitTest[];
-  ClientCertificate: ClientCertificate[];
-  CaCertificate: CaCertificate[];
-  CookieJar: CookieJar[];
-  CloudCredential: CloudProviderCredential[];
-  Settings: Settings[];
-}
+export type { Database } from './models/types';
 
 export const emptyDb = (): Database => ({
   ApiSpec: [],
