+++ packages/insomnia/src/main/window-utils.ts
@@ -25,6 +25,7 @@ import { docsBase } from '../common/documentation';
 import { getElectronStorage } from './electron-storage';
 import { ipcMainOn } from './ipc/electron';
 import { getLogDirectory } from './log';
+import { browserWindows } from './main-window-registry';
 import { createPluginWindow, destroyPluginWindow } from './plugin-window';
 import { MAIN_WINDOW_SECURITY } from './window-security';
 
@@ -32,10 +33,6 @@ const DEFAULT_WIDTH = 1280;
 const DEFAULT_HEIGHT = 720;
 const MINIMUM_WIDTH = 500;
 const MINIMUM_HEIGHT = 400;
-const browserWindows = new Map<'Insomnia' | 'HiddenBrowserWindow', ElectronBrowserWindow>();
-export function getMainWindow(): ElectronBrowserWindow | null {
-  return browserWindows.get('Insomnia') ?? null;
-}
 let hiddenWindowIsBusy = false;
 interface Bounds {
   height?: number;
