+++ packages/insomnia/src/network/network.ts
@@ -10,14 +10,11 @@ import type {
   Request,
   RequestAuthentication,
   RequestGroup,
-  RequestHeader,
   RequestParameter,
   RequestTestResult,
   ResponseTimelineEntry,
   Settings,
-  SocketIORequest,
   UserUploadEnvironment,
-  WebSocketRequest,
   Workspace,
 } from 'insomnia-data';
 import { EnvironmentType, models, services } from 'insomnia-data';
@@ -31,18 +28,20 @@ import { getKVPairFromData } from '~/common/utils/environment-utils';
 import { buildQueryStringFromParams, joinUrlAndQueryString, smartEncodeUrl } from '~/common/utils/url/querystring';
 import { getRuntime } from '~/runtimes';
 
-import type { ExecutionOption, RequestContext } from '../../../insomnia-scripting-environment/src/objects';
-import { SINGLE_VALUE_HEADERS } from '../common/common-headers';
+import type { ExecutionOption } from '../../../insomnia-scripting-environment/src/objects/execution';
+import type { RequestContext } from '../../../insomnia-scripting-environment/src/objects/interfaces';
 import { JSON_ORDER_PREFIX, JSON_ORDER_SEPARATOR } from '../common/constants';
 import { database as db } from '../common/database';
 import { generateId, getContentTypeHeader, getLocationHeader, getSetCookieHeaders } from '../common/misc';
 import { getRenderedRequestAndContext } from '../common/render';
-import { ascendingFirstIndexStringSort } from '../common/sorting';
 import type { ResponsePatch } from '../main/network/libcurl-promise';
 import { QUERY_PARAMS } from './api-key/constants';
-import { getAuthObjectOrNull, isAuthEnabled } from './authentication';
 import { filterClientCertificates } from './certificate';
 import type { TransformedExecuteScriptContext } from './concurrency.renderer';
+import type { sendCurlAndWriteTimelineError, sendCurlAndWriteTimelineResponse } from './network.types';
+
+export { getOrInheritAuthentication, getOrInheritHeaders } from './inherit';
+export type { sendCurlAndWriteTimelineError, sendCurlAndWriteTimelineResponse } from './network.types';
 
 const { isRequest } = models.request;
 const { isRequestGroup } = models.requestGroup;
@@ -51,66 +50,6 @@ export interface SendActionRuntime {
   appendTimeline: (timelinePath: string, logs: string[]) => Promise<void>;
 }
 
-export const getOrInheritAuthentication = ({
-  request,
-  // requestGroups is supposed to be of order leaf to root
-  requestGroups,
-}: {
-  request: Request | WebSocketRequest | SocketIORequest;
-  requestGroups: RequestGroup[];
-}): RequestAuthentication | {} => {
-  const hasValidAuth = getAuthObjectOrNull(request.authentication) && isAuthEnabled(request.authentication);
-  if (hasValidAuth) {
-    return request.authentication;
-  }
-  const hasParentFolders = requestGroups.length > 0;
-  const closestParentFolderWithAuth = requestGroups.find(
-    ({ authentication }) => getAuthObjectOrNull(authentication) && isAuthEnabled(authentication),
-  );
-  const closestAuth = getAuthObjectOrNull(closestParentFolderWithAuth?.authentication);
-  const shouldCheckFolderAuth = hasParentFolders && closestAuth;
-  if (shouldCheckFolderAuth) {
-    // override auth with closest parent folder that has one set
-    return closestAuth;
-  }
-  // if no auth is specified on request or folders, default to none
-  return { type: 'none' };
-};
-export function getOrInheritHeaders({
-  request,
-  requestGroups,
-}: {
-  request: Pick<Request, 'headers'>;
-  requestGroups: Pick<RequestGroup, 'headers'>[];
-}): RequestHeader[] {
-  const httpHeaders = new Map<string, string>();
-  const originalCaseMap = new Map<string, string>();
-  // parent folders, then child folders, then request
-  const headerContexts = [...requestGroups].reverse().concat(request);
-  const headers = headerContexts.flatMap(({ headers }) => headers || []);
-  headers.forEach(({ name, value, disabled }) => {
-    if (disabled || !name.trim()) {
-      return;
-    }
-    const normalizedCase = name.toLowerCase();
-    // preserves the casing of the last header with the same name
-    originalCaseMap.set(normalizedCase, name);
-    const isStrictValueHeader = SINGLE_VALUE_HEADERS.includes(normalizedCase);
-    if (isStrictValueHeader) {
-      httpHeaders.set(normalizedCase, value);
-      return;
-    }
-    // appending will join matching header values with a comma
-    if (httpHeaders.has(normalizedCase)) {
-      httpHeaders.set(normalizedCase, `${httpHeaders.get(normalizedCase)}, ${value}`);
-      return;
-    }
-    httpHeaders.set(normalizedCase, value);
-  });
-  return Array.from(httpHeaders.entries())
-    .sort(ascendingFirstIndexStringSort)
-    .map(([name, value]) => ({ name: originalCaseMap.get(name)!, value }));
-}
 // (only used for getOAuth2 token) Intended to gather all required database objects and initialize ids
 export const fetchRequestGroupData = async (requestGroupId: string) => {
   const requestGroup = await services.requestGroup.getById(requestGroupId);
@@ -789,28 +728,6 @@ export const tryToTransformRequestWithPlugins = async (renderResult: {
   const { request, context } = renderResult;
   return await getRuntime().network.applyRequestHooks(request, context);
 };
-
-export interface sendCurlAndWriteTimelineError {
-  _id: string;
-  parentId: string;
-  timelinePath: string;
-  statusMessage: string;
-  // additional
-  url: string;
-  error: string;
-  elapsedTime: number;
-  bytesRead: number;
-}
-
-export interface sendCurlAndWriteTimelineResponse extends ResponsePatch {
-  _id: string;
-  parentId: string;
-  timelinePath: string;
-  statusMessage: string;
-  cookies: Cookie[];
-  timeline: string[];
-  bytesRead?: number;
-}
 
 export async function sendCurlAndWriteTimeline(
   renderedRequest: RenderedRequest,
