+++ packages/insomnia/src/ui/components/rendered-query-string.tsx
@@ -19,7 +19,7 @@ import { showSettingsModal } from '~/ui/components/modals/settings-modal';
 import { database as db } from '../../common/database';
 import { SECURITY_SETTINGS_PATH_LABEL } from '../../common/misc';
 import { getAuthObjectOrNull, isAuthEnabled } from '../../network/authentication';
-import { getOrInheritAuthentication } from '../../network/network';
+import { getOrInheritAuthentication } from '../../network/inherit';
 import { useNunjucks } from '../context/nunjucks/use-nunjucks';
 import { CopyButton } from './base/copy-button';
