+++ packages/insomnia/src/ui/components/tabs/tab-list.tsx
@@ -26,7 +26,7 @@ import { Icon } from '../icon';
 import { useDocBodyKeyboardShortcuts } from '../keydown-binder';
 import { AddRequestToCollectionModal } from '../modals/add-request-to-collection-modal';
 import { formatMethodName, getRequestMethodShortHand } from '../tags/method-tag';
-import { type BaseTab, InsomniaTab } from './tab';
+import { type BaseTab, InsomniaTab, TAB_CONTEXT_MENU_COMMAND } from './tab';
 
 const { isRequest } = models.request;
 const { isRequestGroup } = models.requestGroup;
@@ -34,11 +34,6 @@ const { isRequestGroup } = models.requestGroup;
 export interface OrganizationTabs {
   tabList: BaseTab[];
   activeTabId?: string;
-}
-
-export const enum TAB_CONTEXT_MENU_COMMAND {
-  CLOSE_ALL = 'Close All',
-  CLOSE_OTHERS = 'Close Other Tabs',
 }
 
 export const OrganizationTabList = ({ showActiveStatus = true, currentPage = '' }) => {
