+++ packages/insomnia/src/ui/components/templating/external-vault/external-vault-form.tsx
@@ -6,7 +6,7 @@ import { Button } from 'react-aria-components';
 import { debounce } from '../../../../common/misc';
 import { Icon } from '../../icon';
 import { CloudCredentialModal } from '../../modals/cloud-credential-modal/cloud-credential-modal';
-import type { ArgConfigFormProps } from '../tag-editor-arg-sub-form';
+import type { ArgConfigFormProps } from '../tag-editor-arg-sub-form.types';
 import { AWSSecretManagerForm } from './aws-secret-manager-form';
 import { AzureKeyVaultForm } from './azure-key-vault-form';
 import { GCPSecretManagerForm } from './gcp-secret-manager-form';
