+++ packages/insomnia/src/ui/components/templating/tag-editor-arg-sub-form.tsx
@@ -1,18 +1,12 @@
-import type { BaseModel } from 'insomnia-data';
 import React from 'react';
 
-import type { NunjucksParsedTag } from '~/common/templating/types';
 import { isBase64String, isValidJSONString } from '~/ui/utils/string-check';
 
 import { ExternalVaultForm } from './external-vault/external-vault-form';
+import type { ArgConfigFormProps } from './tag-editor-arg-sub-form.types';
 
-export interface ArgConfigFormProps {
-  configValue: string;
-  activeTagDefinition: NunjucksParsedTag;
-  activeTagData: NunjucksParsedTag;
-  onChange: (newConfigValue: string) => void;
-  docs: Record<string, BaseModel[]>;
-}
+export type { ArgConfigFormProps } from './tag-editor-arg-sub-form.types';
+
 const formTagNameMapping = {
   vault: ExternalVaultForm,
 };
